feat: add Vision Transformer to the model
This commit adds a Vision Transformer (ViT) to the model.
The ViT is used to encode the visual features of the input image.
This allows the model to better understand the relationship between the image and the text.

The ViT is a transformer-based model that has been shown to achieve state-of-the-art results on a variety of image classification tasks.
It is a powerful model that can be used to extract high-level features from images.

By adding the ViT to the model, we can improve the model's ability to understand the relationship between the image and the text.
This will lead to better performance on tasks such as image captioning and visual question answering.